requirements.txt
Flask
opencv-python
Pillow
numpy
segment-anything
--extra-index-url https://download.pytorch.org/whl/cu118
torch
torchvision